fix: Widget name disappear as you hover over it
